Move utils used by Media only under media.
