+++ assets/src/edit-story/elements/image/output.js
@@ -20,7 +20,7 @@
 import StoryPropTypes from '../../types';
 import { PAGE_WIDTH, PAGE_HEIGHT } from '../../constants';
 import { editorPixels } from '../../units';
-import { getMediaSizePositionProps } from '../shared';
+import { getMediaSizePositionProps } from '../media';
 
 /**
  * Returns AMP HTML for saving into post content for displaying in the FE.
