+++ assets/src/edit-story/elements/media/getFocalFromOffset.js
