+++ assets/src/edit-story/elements/media/getMediaSizePositionProps.js
