+++ assets/src/edit-story/elements/media/scalePanel.js
